Add RetrieveLicense service
This adds the RetrieveLicense service which calls a service to retrieve
all templates for a page and then passes on the result to the
LicenseStore and returning the matched license.

The service is not wired up yet and it's also a first indicator that
using classes and initializing a new client for each service might not
be the best approach. One could have a single client and dependency
inject that one later...