Refactor service structure
This is a refactoring Paul suggested for moving a bit away from too
small (Rails-like) service objects. The idea is to have the domain
better encoded and a less complex hierarchy between the services.

A "service" is now a bigger module encapsulating all steps of a piece of
business logic. Each step call still be called in isolation though.
Technically, the services do not have a single entry point (in fact they
never had if they would include more than one function).

All services are implemented as singletons. And the "core" services make
use of the smaller "util" ones (e.g. client).

For putting together business logic, one simply calls multiple services
in a controller. As the controllers are yet to come, I implemented the
logic in the route directly.