[#13] Implement RetrieveLicense service (#13)
* Move license into models directory

The "license.js" file is not really a service - moving it into a new
"models" directory instead.

Not 100% sure if "models" is the right name, but we'll see ;)

* Update WikiClient tests

Small fix on the wiki client to avoid needing to wrestle around the
linter...

* Add licenses service

Add the licenses service and draft out a call to retrieve the templates
for a page / file from the api and match them against all templates.

* Update wiki url parser

Update the wiki url parser to also be able to parse upload urls. This is
necessary for retrieving the license for a given file url (we have to
query the right api for it, depending on where the file is hosted).

* Update license service

Draft out the happy-path logic for retrieving the license for a file.

* Add a basic unit test for the licenses service

The test for now only covers the happy path.

* Raise error for non-wiki urls

This raises a super generic AssertionError when a file url cannot be
parsed into "title" and "wikiUrl".

* Add FetchOriginalFileData service

Add a service to retrieve the metadata for a file (original url and
authorHtml) for a given title and wikiUrl.

Specific error handling is still missing though...

* Update parseWikiUrl error case

Update the service to already throw an exception of a url cannot be
split into title and wikiUrl instead of returning null.
With returning null, we always have to interfere and check if the return
value is present...

* Return 422 if the original URL cannot be parsed

We now simply forward the error from the URL parser if the new URL is
not a wiki url or just cannot be parsed.

* Return a 404 when no imageinfo can be retrieved

Simply throw an error if the API does not return any useable imageinfo
for a file.

* Add FetchTemplates service

Add a simple service to fetch the templates for a given title and
wikiUrl (so for a given page in the end).

The service is not yet wired up with the rest of the business logic
though.

* Update FetchOriginalFileData service

Update the service to return null for the artistHtml if no Artist can be
found.

* Add RetrieveLicense service

This adds the RetrieveLicense service which calls a service to retrieve
all templates for a page and then passes on the result to the
LicenseStore and returning the matched license.

The service is not wired up yet and it's also a first indicator that
using classes and initializing a new client for each service might not
be the best approach. One could have a single client and dependency
inject that one later...

* Add ParseFileTitle service

This service takes a single file title, verifies the format and then
adds the default commons wikiUrl to it and forwards it to the respective
service to retrieve the additional file data.

* Add ParseIdentifier service

The service takes a url or a title and calls the respective services to
find out the actual wikiUrl and fetch the artistHTML.

* Update License service

Quickly update the license service to make use of the new smaller
services.

* Make client a singleton

This updates all the services to use the same instance of the client. We
construct one in the environment and then inject it everywhere where it
is needed.

* Make LicenseStore a singleton service

We want to have only a single instance of the LicenseStore service (also
because it's rather expensive to build up), therefore we're building it
once and injecting it everywhere where it's necessary.

* Move client and parser into util directory

Since both services act as "helpers" for the actual business logic, I
felt it could be a tidier approach to move them into a dedicated
directory.

* Refactor service structure

This is a refactoring Paul suggested for moving a bit away from too
small (Rails-like) service objects. The idea is to have the domain
better encoded and a less complex hierarchy between the services.

A "service" is now a bigger module encapsulating all steps of a piece of
business logic. Each step call still be called in isolation though.
Technically, the services do not have a single entry point (in fact they
never had if they would include more than one function).

All services are implemented as singletons. And the "core" services make
use of the smaller "util" ones (e.g. client).

For putting together business logic, one simply calls multiple services
in a controller. As the controllers are yet to come, I implemented the
logic in the route directly.

* Add negative for licenses routes

This adds a test checking the error handling for a 404.

* Adapt test structure for compatible license route

Adapt the test structure to fit the new filtered licences route.

* Rename file parameter to fileUrlOrTitle

This renames the parameter of the licences endpoint to better fit the
domain: the content can either be the URL to a file or the prefixed name
of the file (referred to as "title" internally at Wikipedia /
Wikimedia).

* Rename and document fixture file

* Update comment in integration tests

* Update FileData service

Make method async.

* Fix setup and clean up FileData tests

* Make sure to use await when expecting errors

* Add TODOs for marking methods as private

* Refactor method names in License service