Remove the decodeURI step
I found out that the "decodeURI" step we were doing does not actually do
anything - "decodeURIComponent" is what hapi.js uses and what can be
used to decode a URI / a path.

However, we don't need it: the URLs we get from Wikipedia aren't encoded
and the ones we get via the routes are decoded by hapi.js already.