[#14] Add error handling to "/files" endpoint (#16)
* Adapt naming in Files service

* Update Files service public API

Update the Files service to only expose a single method. Aside that: add
some better error handling and the case of returning empty responses.

* Add simple error handling to client

We for now only distinguish between errors while doing a request to an
external API vs. the API not even being reachable.

* Rename url param for files endpoint

Rename the parameter to fit the specification document.

* Update parseWikiUrl helper errors

Update the parseWikiUrl helper to throw a more generic "invalid-url"
error instead of directly throwing "badData" (we want to do that later
in the routes).

* Update error handling in files routes

Update the error handling in the files route to throw a 422 if there's a
problem with parsing the url and otherwise a generic 500 error.

Note: we're not using the error helper any more for that, but explicitly
wrapping and throwing the error.

* Update FileData service public API

Hide the parts we consider private API in the FileData service from
being accessed from the outside.

* Update error handling in licenses route

* Update public API in licenses service

Also: quickly adjust the naming of errors there.

* Remove the generic error function from the server

Since we decided we have to take care of errors directly, we can now
ditch this helper function.

* Add 503 errors to license and files routes

Update the error handling to also include 503 errors when the wiki API is not
reachable.

* Refactor FileData service

Inline the contents of a method and rename the returned attributes
prefixing them with "original" to make it more obvious we're dealing
with the "original file data" (of where the file is hosted / located).

* Update error handling in client

Update the conditional to be negative instead of positive - more
intuitive here.

* Fix typo in tests

* Add separate constant for errors

This adds a constant for all the error strings we're using (and makes
use of the constants).

* Put the custom error handler back in

Put the custom error handler back in, but use it in an adapted, more
simple way for now.

* Add a guard to ensure we only get single pages

This simply adds a check to ensure we're not overlooking the case of
multiple pages in a response from the server.